+++ api/controllers/api/v1/nodeinfo.py
@@ -1,7 +1,6 @@
 from flask import Blueprint, current_app, jsonify, Response, g, abort
 from utils.various import RESTRICTED_NICKNAMES
 from models import db, Config, User, Sound
-from utils.defaults import Reel2bitsDefaults
 
 bp_nodeinfo = Blueprint("bp_nodeinfo", __name__, url_prefix="/nodeinfo")
 
@@ -47,7 +46,7 @@ def nodeinfo(version):
             "nodeDescription": _config.app_description,
             "taxonomy": {"postsName": "Tracks"},
             "restrictedNicknames": RESTRICTED_NICKNAMES,
-            "uploadLimits": {"track": Reel2bitsDefaults.track_size_limit},
+            "uploadLimits": {"track": current_app.config["UPLOAD_TRACK_MAX_SIZE"]},
         },
     }